+++ .gitignore
@@ -1,4 +1,4 @@
 .idea
 node_modules
 dist
-.env
+config/.env
