+++ package.json
@@ -2,7 +2,7 @@
   "name": "youtube-search",
   "version": "1.0.0",
   "description": "EPAM front-end mentoring task about js, public APIs and webpack",
-  "main": "index.js",
+  "main": "app.js",
   "scripts": {
     "dev": "webpack-dev-server --open --config config/webpack.dev.conf.js",
     "build": "webpack --config config/webpack.build.conf.js"
