+++ src/components/finder/finder.scss
@@ -1,4 +1,4 @@
-@import '../utils/vars';
+@import '../../scss/utils/vars';
 
 .search {
   display: flex;
