+++ src/components/slider/slider.scss
@@ -1,4 +1,4 @@
-@import "../utils/vars";
+@import "../../scss/utils/vars";
 
 .slider {
   position: relative;
