+++ src/components/slider/slider_cards.js
