+++ src/components/slider/slider_page.scss
