+++ src/components/slider/slider_paging-control.scss
@@ -1,4 +1,4 @@
-@import '../utils/vars';
+@import '../../scss/utils/vars';
 
 .paging-controls {
   display: flex;
