+++ src/components/spinner/spinner.scss
