+++ src/scss/common/label.scss
